Refactor(SE-59): Adjustment on hero section and footer
